+++ src/pixRequest/pixRequest.php
@@ -35,6 +35,8 @@ class PixRequest extends Resource
     public $cashierType;
     public $tags;
     public $method;
+    public $priority;
+    public $reason;
     public $fee;
     public $status;
     public $flow;
@@ -80,6 +82,8 @@ class PixRequest extends Resource
         -initiatorTaxId [string, default null]: Payment initiator's tax id (CPF/CNPJ). ex: "01234567890" or "20.018.183/0001-80"
         -tags [array of strings, default null]: list of strings for reference when searching for PixRequests. ex: ["employees", "monthly"]
         -method [string, default null]: execution  method for thr creation of the Pix. ex: "manual", "payerQrcode", "dynamicQrcode".
+        -priority [string, default "high"]: specifies the message channel used to send the Pix Request. If set to "high", the request is sent through the primary channel; if set to "low", it uses the secondary channel. Options: "high" or "low"
+        -reason [string, default "customerRequest"]: underlying reason for the payment transaction. ex: "customerRequest", "fraud", "subscriptionFlaw"
 
     ## Attributes (return-only):
         -id [string]: unique id returned when the PixRequest is created. ex: "5656565656565656"
@@ -117,6 +121,8 @@ class PixRequest extends Resource
         $this-> cashierType = Checks::checkParam($params, "cashierType");
         $this-> tags = Checks::checkParam($params, "tags");
         $this-> method = Checks::checkParam($params, "method");
+        $this-> priority = Checks::checkParam($params, "priority");
+        $this-> reason = Checks::checkParam($params, "reason");
         $this-> fee = Checks::checkParam($params, "fee");
         $this-> status = Checks::checkParam($params, "status");
         $this-> flow = Checks::checkParam($params, "flow");